+++ lib/agent_sdk.ml
@@ -108,5 +108,5 @@ let runtime_query = Runtime_query.query
 let query = Query.query
 
 (** Version info *)
-let version = "0.18.0"
+let version = "0.19.0"
 let sdk_name = "agent_sdk"